HBX-1693: Move class 'org.hibernate.tool.hbm2x.TemplateProducer' to 'org.hibernate.tool.internal.export.common.TemplateProducer'
